refactor: improve parsing performance
Profiling showed that a *lot* time was spent in object parsing, mainly
with iterating over the object declaration/object data list trying to
find an object by ID with a specific partition ID.

Now we build a (object ID, partition ID) to object data hash map in the
object group parser so we don't spend all this time traversing a list
but can use a hash lookup to find the object data.

This was also in a sort of hot code path as an object group may contain
quite a few objects for which the object parsing code is called.