fix: correctly skip older object versions when parsing base revisions
When parsing an object space we start with the most recent revision
specified by the storage index. We first parse all objects in this
revision and then proceed through the list of base revisions, parsing
all objects that are not contained within more recent revisions.

Only that we didn't. Instead of checking if the object already exists,
we erroneously checked if the object *group* already exists as a parsed
object - which it never did.

This is now fixed :)

This showed by some page contents being *very* old while some more
recent contents where present too. Basically this updated the object
space to always contain the earliest possible version of every object.