refactor: use a cache to not parse revisions multiple times
In a OneNote an object space may point to a revision that is a base
revision of another object space. Naively parsing a revisions base
revisions may thus result in some revisions being parsed multiple times.

This introduces a revision cache where parsed revisions are stored. If
we find that we're asked to parse a revision we've parsed before, we
just copy all objects from the cache into the object space.