refactor: use stdlib hashmap instead of index map
IndexMap seems to be optimized for a case where accessing data by index
or iterating in order is a common operation. In our case we only need
to know the insertion order in order to calculate the object id offset
for parsing object references (as well as object space id/references).
And even then we don't care about the insertion order per se but rather
need to count all object references *before* the current property.

By moving the insertion order index into the hash map value insertion
itself stays as cheap as possible and we move the cost of looking up
the index to where we actually need to know it - when calculating the
reference offset.