fix: correctly parse boolean properties
Boolean properties have their values contained within the property ID.
For that reason we can't look up boolean properties by their property ID
as we'd only find property set entries where the value is `false`.
Instead we calculate the ID part of the property ID and use this to find
the desired values.