fix: properly parse base revisions
When parsing a base revision, we only want to keep objects in the object
space that are not part of newer revisions. But as it was up till now,
we parsed the base revision and added ALL objects into the object space
instead of only adding new objects. This fixes this wrong behavior.