fix: handle data elements with same ID but different type
In some OneNote files there are multiple data elements with the
same ID but with different types. Up till now this lead to issues as
in this case only one data element was kept while others were
discarded.

With this change we store data elements by type in separate hash maps
so we can correctly handle this situation.