PropertType: Specify type for custom discriminant values
We use custom discriminant values for enum variants for "PropertType" without specifying the type.
This causes it to default to "isize" which on 64bit platforms is big enough to hold the 32bit constants,
but causes overflow when isize is 32bit.

This issue was uncovered due to build failures on arm for Gentoo Linux as a dependency of ClamAV.

Bug: https://bugs.gentoo.org/927214
Signed-off-by: Matt Jolly <kangie@gentoo.org>
Co-authored-by: Steven Newbury <steve@snewbury.org.uk>